apps/workers/dor_processor/requirements.txt
duckdb>=0.9.0
polars>=0.20.0
pyarrow>=14.0.0
pandas>=2.1.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
aiohttp>=3.9.0
asyncpg>=0.29.0
supabase>=2.3.0
watchdog>=3.0.0
schedule==1.2.0
python-dotenv>=1.0.0